requirements.txt
django>=4.0.0,<4.2.0
djangorestframework
pyyaml
requests
django-cors-headers
Pillow
django-phonenumber-field[phonenumbers]